+++ CHANGELOG.md
@@ -84,7 +84,7 @@
 - Upgrade to TypeScript 5.7
 - Upgraded dependencies
   - Now supports React 19
-  - `effect` (needs to be 3.12.0 or higher)
+  - `effect` (needs to be 3.13.0 or higher)
   - `@livestore/wa-sqlite` (needs to be 1.0.3-dev.7)
 
 ## 0.2.0